Data/Agent/agent-requirements.txt
# API / WebSocket Handling
requests
python-socketio
websocket-client
eventlet
aiohttp

# GUI-related dependencies (Qt for GUI components)
PyQt5
qasync

# Computer Vision & OCR Dependencies
Pillow